feat: Add lesson on gridworld
Lesson 9 continues reinforcement learning with an example
with the gridworld problem. The random movement policy is
evaluated in a 2x2 gridworld, and the state value function
is computed iteratively. The gridworld can be visualized at
any step in the iteration. The greedy policy for any step in
the iteration can also be visualized.